test: CarServiceImplTest 삭제 후 Car Name 검증 기능을 ModelTest로 분배
정적 팩토리 메서드를 사용하는 Car과 Participation class의 생성에서 검증을 테스트함.
